+++ gdsfactory/plugins/database/sql_base.py
@@ -1,5 +1,0 @@
-from sqlalchemy.ext.declarative import declarative_base
-
-# this avoids circular imports for having models in many files
-Base = declarative_base()
-metadata = Base.metadata